Split RpcEventProducer and ReconnectingRpcEventProducer files
